django-worker/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
coverage==5.5
cycler==0.10.0
Django==3.2.6
djangorestframework==3.12.4
Faker==8.14.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hypothesis==6.23.0
iniconfig==1.1.1
jedi==0.18.0
kiwisolver==1.3.2
#matplotlib==3.4.3
mixer==7.1.2
numpy==1.21.2
packaging==21.0
parso==0.8.2
Pillow==8.3.2
pluggy==1.0.0
promise==2.3
#psycopg2-binary==2.9.1
pudb==2021.1
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urwid==2.1.2